Refactor user invitation components by reorganizing namespaces and adding GetAllUserInvitationsInput DTO
