+++ aspnet-core/src/toyiyo.todo.Application/Invitations/UserInvitationService.cs
@@ -1,12 +1,11 @@
 using System.Threading.Tasks;
 using Abp.Authorization;
 using toyiyo.todo.Authorization;
-using toyiyo.todo.Authorization.Users;
 using toyiyo.todo.Invitations.Dto;
 
 namespace toyiyo.todo.Invitations
 {
-     [AbpAuthorize(PermissionNames.Pages_Subscription)]
+    [AbpAuthorize(PermissionNames.Pages_Subscription)]
     public class UserInvitationAppService : todoAppServiceBase, IUserInvitationAppService
     {
         private readonly IUserInvitationManager _userInvitationManager;
