+++ aspnet-core/src/toyiyo.todo.Core/Invitations/GetAllUserInvitationsInput.cs
@@ -1,0 +1,9 @@
+using Abp.Application.Services.Dto;
+
+namespace toyiyo.todo.Invitations
+{
+    public class GetAllUserInvitationsInput : PagedAndSortedResultRequestDto
+    {
+        public string Keyword { get; set; } 
+    }
+}
